MyApps-13: Move waves implementation, api and dependency injection to submodules.
