+++ app/src/main/cpp/CMakeLists.txt
@@ -6,7 +6,6 @@
 cmake_minimum_required(VERSION 3.22.1)
 
 add_subdirectory(basic)
-add_subdirectory(waves)
 add_subdirectory(myengine)
 add_subdirectory(web)
 add_subdirectory(calculator)
