+++ app/src/main/java/xyz/torquato/myapps/di/data/DataModule.kt
@@ -5,13 +5,11 @@ import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import xyz.torquato.myapps.api.math.IMathRepository
-import xyz.torquato.myapps.api.sound.ISoundRepository
 import xyz.torquato.myapps.api.web.ILocalWebRepository
 import xyz.torquato.myapps.api.web.IQueryRepository
 import xyz.torquato.myapps.api.web.IWebRepository
 import xyz.torquato.myapps.api.web.cache.IWebCacheRepository
 import xyz.torquato.myapps.data.math.MathRepository
-import xyz.torquato.myapps.data.waves.SoundRepository
 import xyz.torquato.myapps.data.web.LocalWebRepository
 import xyz.torquato.myapps.data.web.QueryRepository
 import xyz.torquato.myapps.data.web.WebRepository
@@ -21,11 +19,6 @@ import javax.inject.Singleton
 @Module
 @InstallIn(SingletonComponent::class)
 abstract class RepositoryModule {
-
-    @Binds
-    abstract fun bindImpl(impl: SoundRepository): ISoundRepository
-
-
 
     @Binds
     @Singleton
