+++ data/impl/src/main/cpp/CMakeLists.txt
@@ -6,3 +6,4 @@
 cmake_minimum_required(VERSION 3.22.1)
 
 add_subdirectory(cypher)
+add_subdirectory(waves)
