+++ data/impl/src/main/cpp/waves/api/Oscillator.h
