+++ data/impl/src/main/cpp/waves/api/model/Tone.h
